refactor: organize card components
