+++ website-frontend/src/lib/components/cards/PeopleCard.svelte
