+++ website-frontend/src/lib/components/cards/PublicationCard.svelte
