+++ website-frontend/src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	/** @type {import('./$types').PageData} */
 	import Hero from '$lib/components/carousels/LandingHero.svelte';
-	import FeaturedEventCard from '$lib/components/events/FeaturedEventCard.svelte';
+	import FeaturedEventCard from '$lib/components/cards/FeaturedEventCard.svelte';
 
 	export let data;
