+++ website-frontend/src/routes/events/+page.svelte
@@ -2,7 +2,7 @@
 	/** @type {import('./$types').PageData} */
 	import Banner from '$lib/components/banner/Banner.svelte';
 	import * as Carousel from '$lib/@shadcn-svelte/ui/carousel/index';
-	import FeaturedEventCard from '$lib/components/events/FeaturedEventCard.svelte';
+	import FeaturedEventCard from '$lib/components/cards/FeaturedEventCard.svelte';
 	import FilterBar from '$lib/components/filter/FilterBar.svelte';
 	import LoadMore from '$lib/components/buttons/LoadMore.svelte';
