+++ website-frontend/src/routes/research/outputs/+page.svelte
@@ -5,7 +5,7 @@
 	import CardPanel from '$lib/components/panel/CardPanel.svelte';
 	import FilterBar from '$lib/components/filter/FilterBar.svelte';
 	import type { FilterControls } from '$lib/types/filter_controls.js';
-	import PublicationCard from '$lib/components/card/PublicationCard.svelte';
+	import PublicationCard from '$lib/components/cards/PublicationCard.svelte';
 
 	export let data;
